+++ wsgidav/dir_browser_2/__init__.py
@@ -1,2 +1,0 @@
-# flake8: noqa
-from ._dir_browser import WsgiDavDirBrowser2